poco update to 1.4.6

+++ trunk/public/include/Poco/Thread_POSIX.h
@@ -1,7 +1,7 @@
 //
 // Thread_POSIX.h
 //
-// $Id: //poco/1.4/Foundation/include/Poco/Thread_POSIX.h#5 $
+// $Id: //poco/1.4/Foundation/include/Poco/Thread_POSIX.h#6 $
 //
 // Library: Foundation
 // Package: Threading
@@ -215,7 +215,7 @@ inline void ThreadImpl::yieldImpl()
 
 inline int ThreadImpl::getStackSizeImpl() const
 {
-	return _pData->stackSize;
+	return static_cast<int>(_pData->stackSize);
 }
